+++ src/main/java/me/desair/tus/server/download/DownloadGetRequestHandler.java
@@ -57,12 +57,11 @@ public class DownloadGetRequestHandler extends AbstractRequestHandler {
           HttpHeader.CONTENT_DISPOSITION,
           String.format(
               CONTENT_DISPOSITION_FORMAT,
-              info.getFileName().replaceAll("[\r\n\"]", ""),
+              info.getFileName().replace("\"", ""),
               URLEncoder.encode(info.getFileName(), StandardCharsets.UTF_8.toString())
                   .replace("+", "%20")));
 
-      servletResponse.setHeader(
-          HttpHeader.CONTENT_TYPE, info.getFileMimeType().replaceAll("[\r\n]", ""));
+      servletResponse.setHeader(HttpHeader.CONTENT_TYPE, info.getFileMimeType());
 
       uploadStorageService.copyUploadTo(info, servletResponse.getOutputStream());
     }