+++ src/test/java/me/desair/tus/server/creationwithupload/validation/UploadContentLengthValidatorTest.java
@@ -1,4 +1,4 @@
-package me.desair.tus.server.creation.validation;
+package me.desair.tus.server.creationwithupload.validation;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -11,16 +11,15 @@ import org.junit.Before;
 import org.junit.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 
-public class PostEmptyRequestValidatorTest {
+public class UploadContentLengthValidatorTest {
 
-  private PostEmptyRequestValidator validator;
-
+  private UploadContentLengthValidator validator;
   private MockHttpServletRequest servletRequest;
 
   @Before
   public void setUp() {
     servletRequest = new MockHttpServletRequest();
-    validator = new PostEmptyRequestValidator();
+    validator = new UploadContentLengthValidator();
   }
 
   @Test
@@ -36,56 +35,56 @@ public class PostEmptyRequestValidatorTest {
   }
 
   @Test
-  public void validateMissingContentLength() throws Exception {
-    // We don't set a content length header
-    // servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 3L);
+  public void validateContentLengthLessOrEqualUploadLength() throws Exception {
+    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 100L);
+    servletRequest.addHeader(HttpHeader.UPLOAD_LENGTH, 150L);
 
-    // When we validate the request
     try {
       validator.validate(HttpMethod.POST, servletRequest, null, null);
     } catch (Exception ex) {
       fail();
     }
-
-    // No Exception is thrown
   }
 
   @Test
-  public void validateContentLengthZero() throws Exception {
-    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 0L);
+  public void validateContentLengthEqualUploadLength() throws Exception {
+    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 100L);
+    servletRequest.addHeader(HttpHeader.UPLOAD_LENGTH, 100L);
 
-    // When we validate the request
     try {
       validator.validate(HttpMethod.POST, servletRequest, null, null);
     } catch (Exception ex) {
       fail();
     }
-
-    // No Exception is thrown
   }
 
   @Test(expected = InvalidContentLengthException.class)
-  public void validateContentLengthNotZero() throws Exception {
-    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 10L);
+  public void validateContentLengthExceedsUploadLength() throws Exception {
+    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 150L);
+    servletRequest.addHeader(HttpHeader.UPLOAD_LENGTH, 100L);
 
-    // When we validate the request
     validator.validate(HttpMethod.POST, servletRequest, null, null);
-
-    // Expect a InvalidContentLengthException
   }
 
   @Test
-  public void validateContentLengthNotZeroWithCreationWithUpload() throws Exception {
-    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 10L);
-    validator.setCreationWithUploadEnabled(true);
+  public void validateNoContentLength() throws Exception {
+    servletRequest.addHeader(HttpHeader.UPLOAD_LENGTH, 100L);
 
-    // When we validate the request
     try {
       validator.validate(HttpMethod.POST, servletRequest, null, null);
     } catch (Exception ex) {
       fail();
     }
+  }
 
-    // No Exception is thrown
+  @Test
+  public void validateNoUploadLength() throws Exception {
+    servletRequest.addHeader(HttpHeader.CONTENT_LENGTH, 100L);
+
+    try {
+      validator.validate(HttpMethod.POST, servletRequest, null, null);
+    } catch (Exception ex) {
+      fail();
+    }
   }
 }